+++ apx_sdk/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Added `ed25519_public_key_to_pkcs8_pem` function.
+
 ### Changed
 
 - Add response URL to `FetchError::JsonParseError`.